refactor: order와 table 분리
