+++ src/main/java/kitchenpos/table/domain/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.domain.repository;
+package kitchenpos.table.domain;
 
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
