+++ src/test/java/kitchenpos/order/application/TableServiceTest.java
@@ -1,12 +1,12 @@
 package kitchenpos.order.application;
 
-import kitchenpos.order.domain.*;
-import kitchenpos.order.application.TableService;
+import kitchenpos.table.application.TableService;
 import kitchenpos.order.domain.repository.OrderRepository;
-import kitchenpos.order.domain.repository.OrderTableRepository;
-import kitchenpos.order.dto.OrderTableRequest;
-import kitchenpos.order.dto.OrderTableResponse;
-import org.junit.jupiter.api.BeforeEach;
+import kitchenpos.table.domain.OrderTableRepository;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
