+++ src/test/java/kitchenpos/order/event/OrderLineItemsEventTest.java
@@ -3,8 +3,8 @@ package kitchenpos.order.event;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.*;
 import kitchenpos.order.domain.repository.OrderLineItemRepository;
-import kitchenpos.order.event.order.OrderCreatedEvent;
-import kitchenpos.order.event.order.OrderLineItemsEventHandler;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
